REVERB

+++ sound_generator.h
@@ -35,7 +35,7 @@ int add_delay_soundgen(SOUNDGEN *self, float duration);
 int add_moddelay_soundgen(SOUNDGEN *self);
 int add_reverb_soundgen(SOUNDGEN *self);
 int add_freq_pass_soundgen(SOUNDGEN *self, float freq, effect_type pass_type);
-float effector(SOUNDGEN *self, float val);
+float effector(SOUNDGEN *self, double val);
 
 int add_envelope_soundgen(SOUNDGEN *self, ENVSTREAM *e);
 float envelopor(SOUNDGEN *self, float val);
